refactor: summation ensemble files to one file remove duplicated code
